Only install dockutil if it is not already installed.

+++ scripts/common/configuration-osx.sh
@@ -21,9 +21,11 @@ defaults write com.apple.finder '_FXShowPosixPathInTitle' -bool true
 defaults -currentHost write com.apple.ImageCapture disableHotPlug -bool true
 #to revert use defaults -currentHost delete com.apple.ImageCapture disableHotPlug
 
-
 # modify appearance of dock: remove standard icons, add chrome and iTerm
-brew install dockutil
+if [ ! -e /usr/local/bin/dockutil ]; then
+  # dockutil is not installed
+  brew install dockutil
+fi
 dockutil --list | awk -F\t '{print "dockutil --remove \""$1"\" --no-restart"}' | sh
 dockutil --add /Applications/Google\ Chrome.app --no-restart
 dockutil --add /Applications/iTerm.app
